pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code-auditor"
version = "0.1.0"
description = "Multi-stage code auditing agent application using Claude Code or Codex agent backends"
requires-python = ">=3.12"
dependencies = [
    "claude-code-sdk>=0.0.25",
]

[project.optional-dependencies]
codex = [
    "codex-app-server-sdk @ git+https://github.com/openai/codex.git@main#subdirectory=sdk/python",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.0",
]

[project.scripts]
code-auditor = "code_auditor.__main__:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
